fix(mcp): extract role names as strings in UserInfo serialization
UserInfo.model_validate(owner, from_attributes=True) passed raw
SQLAlchemy Role objects to the roles field which expects List[str],
causing Pydantic validation errors in get_dashboard_info and similar
tools.

Move serialize_user_object to system/schemas.py and use it in chart,
dashboard, and dataset serializers to properly extract role.name
strings from ORM Role objects.
